# public/openapi.json
{
  "info": {
    "title": "Users API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "listUsers",
        "summary": "List users",
        "description": "List all users",
        "responses": {
          "default": {
            "description": "",
            "status": "default",
            "mediaType": "application/json",
            "content": {
              "application/json": {}
            }
          }
        },
        "parameters": [],
        "security": [
          {}
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "getUserById",
        "summary": "Get user by id",
        "description": "Get user by identifier",
        "responses": {
          "default": {
            "description": "",
            "status": "default",
            "mediaType": "application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {}
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "createdAt",
          "isActive"
        ]
      }
    }
  }
}